TeamAnalyzer: Updated to 2.70, changed "Diff%" to "Relative" and using (myRating/(myRating+opponentRating)) as new formula, closes #78
git-svn-id: file:///home/coren/project/backup_ho/trunk@517 b8de521a-fa13-0410-8113-fc7ffabbd59a

+++ Plugin-TrainingExperience/src/java/hoplugins/trainingExperience/ui/bar/VerticalIndicator.java
@@ -15,8 +15,8 @@ import javax.swing.JPanel;
 public class VerticalIndicator extends JPanel {
     //~ Instance fields ----------------------------------------------------------------------------
 
-    private float actual;
-    private float total;
+    private double actual;
+    private double total;
 
     //~ Constructors -------------------------------------------------------------------------------
 
@@ -26,7 +26,7 @@ public class VerticalIndicator extends JPanel {
      * @param _actual actual points
      * @param _total skillup points
      */
-    public VerticalIndicator(float _actual, float _total) {
+    public VerticalIndicator(double _actual, double _total) {
         super();
         setLayout(new BorderLayout());
         setOpaque(false);
